refactor: rename directory for ts modules
