+++ src/popup/components/Views/GroupView/GroupControls/MenuItems/ExportGroupMenuItem.vue
@@ -4,7 +4,7 @@ import { useGroupStore } from '@/stores/group'
 import { usePopupStore } from '@/stores/popup'
 import { trans } from '@common/modules/trans'
 import { error } from '@common/modules/error'
-import { encryptGroup } from '@common/modules/encrypt/encryptGroup'
+import { encryptGroup } from '@common/modules/crypto/encryptGroup'
 import ArrowDownTrayIcon from '@common/components/Icons/ArrowDownTrayIcon.vue'
 import MenuItem from '@/components/MenuItem.vue'
 import { getPasswordFromStorage } from '@common/modules/storage/getPasswordFromStorage'
