+++ src/popup/stores/group.ts
@@ -11,9 +11,9 @@ import { isIncognito } from '@common/modules/browser/windows'
 import { deleteGroupFromStorage } from '@common/modules/storage/deleteGroupFromStorage'
 import { getGroupsFromStorage } from '@common/modules/storage/getGroupsFromStorage'
 import { getDefaultGroupName } from '@/modules/getDefaultGroupName'
-import { decryptGroup as decryptGroupAlias } from '@common/modules/encrypt/decryptGroup'
 import { saveGroupToStorage } from '@common/modules/storage/saveGroupToStorage'
-import { encryptGroup } from '@common/modules/encrypt/encryptGroup'
+import { decryptGroup as decryptGroupAlias } from '@common/modules/crypto/decryptGroup'
+import { encryptGroup } from '@common/modules/crypto/encryptGroup'
 import { closeTabsByIds } from '@/modules/tabs/closeTabsByIds'
 import { getCurrentURL } from '@/modules/getCurrentURL'
 import { deleteAllGroupsFromStorage } from '@common/modules/storage/deleteAllGroupsFromStorage'
